requirements.txt
Flask
flask-cors
SQLAlchemy
pymysql
python-dotenv
